requirements.txt
# requirements.txt
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
bcrypt>=4.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
SpeechRecognition>=3.10.0
scipy>=1.11.0
pyobjc-framework-LocalAuthentication>=9.0; sys_platform == 'darwin'